requirements.txt
streamlit>=1.32.0
yfinance>=0.2.37
pandas>=2.0.0
plotly>=5.20.0
python-telegram-bot>=20.0
kaleido>=0.2.1
